pyproject.toml
[tool.poetry]
name = "cleanab"
version = "0.1.0"
description = "Fetch and clean FinTS/HBCI transactions and push them to YNAB"
authors = ["Jan Willhaus <mail@janwillhaus.de>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.7"
fints = "^3.0"
pyyaml = "^5.1"
logzero = "^1.5.0"
click = "^7.1.2"
ynab-api = {git = "https://github.com/dmlerner/ynab-api"}
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
pycleanab = 'cleanab.main:main'

[tool.isort]
profile = "black"
multi_line_output = 3

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
